modify install log.info messages and set command as a class
