+++ nf_core/components/update.py
@@ -4,10 +4,9 @@ import shutil
 import tempfile
 from pathlib import Path
 
-import questionary
-
 import nf_core.modules.modules_utils
 import nf_core.utils
+import questionary
 from nf_core.components.components_command import ComponentCommand
 from nf_core.components.components_utils import prompt_component_version_sha
 from nf_core.modules.modules_differ import ModulesDiffer
@@ -248,7 +247,7 @@ class ComponentUpdate(ComponentCommand):
                                 style=nf_core.utils.nfcore_question_style,
                             ).unsafe_ask()
                     if recursive_update and len(modules_to_update + subworkflows_to_update) > 0:
-                        # Write all the differences of linked componenets to a diff file
+                        # Write all the differences of linked components to a diff file
                         self.update_linked_components(
                             modules_to_update, subworkflows_to_update, updated, check_diff_exist=False
                         )
